rename: similarity 패키지 오탈자 수정 #118
