+++ src/main/java/kr/co/csalgo/config/FeedbackAnalyzerConfig.java
@@ -4,7 +4,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import kr.co.csalgo.domain.question.feedback.FeedbackAnalyzer;
-import kr.co.csalgo.domain.silmilarity.SimilarityCalculator;
+import kr.co.csalgo.domain.similarity.SimilarityCalculator;
 import kr.co.csalgo.infrastructure.feedback.SimpleFeedbackAnalyzer;
 import lombok.RequiredArgsConstructor;
