+++ src/main/java/kr/co/csalgo/config/SimilarityConfig.java
@@ -3,7 +3,7 @@ package kr.co.csalgo.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import kr.co.csalgo.domain.silmilarity.SimilarityCalculator;
+import kr.co.csalgo.domain.similarity.SimilarityCalculator;
 import kr.co.csalgo.infrastructure.similarity.LuceneSimilarityCalculator;
 
 @Configuration
