+++ src/main/java/kr/co/csalgo/domain/similarity/SimilarityCalculator.java
@@ -1,4 +1,4 @@
-package kr.co.csalgo.domain.silmilarity;
+package kr.co.csalgo.domain.similarity;
 
 public interface SimilarityCalculator {
 	double calculate(String reference, String userInput);
