+++ src/main/java/kr/co/csalgo/domain/similarity/SimilarityGuide.java
@@ -1,4 +1,4 @@
-package kr.co.csalgo.domain.silmilarity;
+package kr.co.csalgo.domain.similarity;
 
 import lombok.Getter;
